533 Move rwcds-form optionsets into their own file
