+++ client/app/optionsets/rwcds-form.js
@@ -1,0 +1,47 @@
+export const DCPHASPROJECTCHANGEDSINCESUBMISSIONOFTHEPAS_OPTIONSET = {
+  YES: {
+    code: true,
+    label: 'Yes',
+  },
+  NO: {
+    code: false,
+    label: 'No',
+  },
+};
+
+export const DCPCONSTRUCTIONPHASING_OPTIONSET = {
+  SINGLE: {
+    code: 717170000,
+    label: 'Single',
+  },
+  MULTIPLE: {
+    code: 717170001,
+    label: 'Multiple',
+  },
+  NOT_APPLICABLE: {
+    code: null,
+    label: 'Not Applicable',
+  },
+};
+
+export const DCPEXISTINGCONDITIONS_OPTIONSET = {
+  YES: {
+    code: true,
+    label: 'Yes',
+  },
+  NO: {
+    code: false,
+    label: 'No',
+  },
+};
+
+export const DCPISRWCDSSCENARIO_OPTIONSET = {
+  YES: {
+    code: true,
+    label: 'Yes',
+  },
+  NO: {
+    code: false,
+    label: 'No',
+  },
+};
